Remove optimization that most likely caused unstable tests
The intention was that it was unnecessary to queue up additional reads
if there was already a task waiting to read, but the early return could
cause tests to fail because the intentional call from the tests to read
messages sometimes did nothing and the test continued before the
messages were read.

In theory this could case the file to be read more times than necessary
but in real world use I expect this to be an edge case so it shouldn't
affect performance much at all.