Attempt to stabilize initialization race conditions
Because TinyIpc utilizes background worker tasks that sometimes can take
a little while to start the tests have been somewhat unstable on slower
machines such as the GitHub build agents.

It can be argued that the services are not quite ready when the
constructor is done but only ready when the worker tasks have also
started.

Because of this I have added task completion sources that are awaited
when the services are interacted with in ways where the worker task is
expected to be running. This should improve test stability and also make
TinyIpc more predictable.